fix: accumulate the MaxSim sum in fp32 to preserve retrieval quality
colbert_scores summed the per-query-token maxima in the input dtype:

    scores = scores.max(axis=-1).values.sum(axis=-1)

Under fp16 or bf16 that runs in reduced precision across ~32 query tokens
and costs measurable retrieval quality.

Measured on BEIR SciFact -- 1,500 documents (every judged document plus
distractors), 300 queries, lightonai/GTE-ModernColBERT-v1, embeddings
computed once and cached so the accumulation is the only variable:

    accumulation                nDCG@10    Recall@10
    fp32 (reference)            0.854361   0.937000
    fp16, sum in fp16           0.850314   0.927000
    fp16, sum in fp32           0.854400   0.937000
    bf16, sum in bf16           0.832469   0.918667
    bf16, sum in fp32           0.851472   0.927000

bf16 loses 2.2 nDCG points and 1.8 points of Recall@10; summing in fp32
recovers essentially all of it. This is the documented path rather than a
corner case: fp16=True appears in docs/index.md, docs/documentation/
training.md and two example training scripts, bf16=True in two more, and
instrumenting Contrastive -> ColBERTScores -> colbert_scores shows the
loss receives fp16/bf16 under the corresponding autocast.

It also explains a backend discrepancy: with late-interaction-kernels
installed, backend="auto" scores fp16 at 0.854361, matching fp32, because
the LIK kernel already accumulates in fp32. Without it, the same call
returns 0.850314.

Note this makes colbert_scores return fp32 rather than the input dtype.

The added test fails on the unpatched source for both fp16 and bf16.

Measured on macOS arm64 (torch 2.11.0, sentence-transformers 5.3.0); I
have no CUDA hardware, so the numbers are not confirmed there, though the
mechanism is device-independent.

+++ pylate/scores/colbert.py
@@ -197,7 +197,9 @@ def colbert_scores(
 
     if documents_mask is not None:
         scores = scores * documents_mask.unsqueeze(0).unsqueeze(2)
-    scores = scores.max(axis=-1).values.sum(axis=-1)
+    # accumulate the per-query-token maxima in fp32: summing ~32 terms in fp16/bf16
+    # loses retrieval quality, and reduced precision is the documented training path
+    scores = scores.max(axis=-1).values.sum(axis=-1, dtype=torch.float32)
     return scores
